fix: gate GET /api/applications with admin session
job_applications PII was readable with INDEXER_API_KEY via requireAuth. Use Supabase session and admin_users like admin bug reports.

Extract requireAdminSession to lib/admin-session.ts and reuse in admin/bugs.

Made-with: Cursor
