+++ app/lib/admin-session.ts
@@ -1,0 +1,65 @@
+import { createServerClient } from "@supabase/ssr";
+import { cookies } from "next/headers";
+import type { User } from "@supabase/supabase-js";
+import { NextResponse } from "next/server";
+import { getServiceClient } from "@/lib/supabase";
+
+export type AdminSessionResult =
+  | { ok: true; user: User }
+  | { ok: false; response: NextResponse };
+
+/**
+ * Verify Supabase Auth session (cookies) and membership in `admin_users`.
+ * For Route Handlers that return PII using `getServiceClient()`.
+ */
+export async function requireAdminSession(): Promise<AdminSessionResult> {
+  const cookieStore = await cookies();
+  const supabase = createServerClient(
+    process.env.NEXT_PUBLIC_SUPABASE_URL!,
+    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
+    {
+      cookies: {
+        getAll() {
+          return cookieStore.getAll();
+        },
+        setAll() {
+          /* read-only in Route Handlers */
+        },
+      },
+    },
+  );
+
+  const {
+    data: { user },
+    error,
+  } = await supabase.auth.getUser();
+  if (error || !user) {
+    return {
+      ok: false,
+      response: NextResponse.json({ error: "Unauthorized" }, { status: 401 }),
+    };
+  }
+  if (!user.email) {
+    return {
+      ok: false,
+      response: NextResponse.json({ error: "Unauthorized" }, { status: 401 }),
+    };
+  }
+
+  const sb = getServiceClient();
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const { data: adminRow } = await (sb as any)
+    .from("admin_users")
+    .select("id")
+    .eq("email", user.email)
+    .maybeSingle();
+
+  if (!adminRow) {
+    return {
+      ok: false,
+      response: NextResponse.json({ error: "Forbidden" }, { status: 403 }),
+    };
+  }
+
+  return { ok: true, user };
+}
